feat(backend): inserting test
New integration and unit test

+++ backend/src/controllers/incidentcontroller.js
@@ -41,10 +41,13 @@ module.exports = {
     const ong_id = request.headers.authorization;
     const incident = await connection('incidents').where('id', id).select('ong_id').first();
 
-    if (incident.ong_id !== ong_id) {
-      return response.status(401).json({ error: 'Operation not permitted' });
+    if (incident) {
+      // return response.status(404).json({ error: 'Not found' });
+      if (incident.ong_id !== ong_id) {
+        return response.status(401).json({ error: 'Operation not permitted' });
+      }
+      await connection('incidents').where('id', id).delete();
     }
-    await connection('incidents').where('id', id).delete();
     return response.status(204).send();
   },
 };
